Set up annotations & methods in controllers + assert on entities attributes + adding slug attribute to Trick entity + removing style tags to my Twig files + set up the navbar in a different Twig file
