+++ templates/figures/index.html.twig
@@ -1,25 +1,22 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Accueil{% endblock %}
+{% block title %}Accueil
+{% endblock %}
 
 {% block body %}
-<style>
-    .example-wrapper { margin: 1em auto; max-width: 800px; width: 95%; font: 18px/1.5 sans-serif; }
-    .example-wrapper code { background: #F5F5F5; padding: 2px 6px; }
-</style>
 
-<div class="example-wrapper">
-    <section class="figures">
-        <article>
-            <h2>Nom de la figure</h2>
-            <div class="metadata">Rédigé le 19/11/2021 à 15h00 dans la catégorie Figures</div>
-            <div class="content">
-                <img src="http://via.placeholder.com/350x150" alt="">
-                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reprehenderit facere illo nulla dolorem. Illo, ab.</p>
-                <p>Consequatur necessitatibus nostrum eaque. Omnis dolorum unde facilis eligendi odio quos quaerat debitis non quo!</p>
-                <a href="{{ path('figures_show') }}" class="btn btn-primary">Lire la suite</a>
-            </div>
-        </article>
-    </section>
-</div>
+	<div class="example-wrapper">
+		<section class="figures">
+			<article>
+				<h2>Nom de la figure</h2>
+				<div class="metadata">Rédigé le 19/11/2021 à 15h00 dans la catégorie Figures</div>
+				<div class="content">
+					<img src="http://via.placeholder.com/350x150" alt="">
+					<p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reprehenderit facere illo nulla dolorem. Illo, ab.</p>
+					<p>Consequatur necessitatibus nostrum eaque. Omnis dolorum unde facilis eligendi odio quos quaerat debitis non quo!</p>
+					<a href="{{ path('figures_show') }}" class="btn btn-primary">Lire la suite</a>
+				</div>
+			</article>
+		</section>
+	</div>
 {% endblock %}
