+++ templates/figures/show.html.twig
@@ -1,27 +1,24 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Accueil{% endblock %}
+{% block title %}Accueil
+{% endblock %}
 
 {% block body %}
-<style>
-    .example-wrapper { margin: 1em auto; max-width: 800px; width: 95%; font: 18px/1.5 sans-serif; }
-    .example-wrapper code { background: #F5F5F5; padding: 2px 6px; }
-</style>
 
-<div class="example-wrapper">
-    <section class="figures">
-        <article>
-            <h2>Nom de la figure</h2>
-            <div class="metadata">Rédigé le 19/11/2021 à 15h00 dans la catégorie Figures</div>
-            <div class="content">
-                <img src="http://via.placeholder.com/350x150" alt="">
-                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reprehenderit facere illo nulla dolorem. Illo, ab.</p>
-                <p>Consequatur necessitatibus nostrum eaque. Omnis dolorum unde facilis eligendi odio quos quaerat debitis non quo!</p>
-                <hr>
-                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Itaque quo perspiciatis vero. Aliquam itaque aliquid, nostrum est dolor excepturi omnis blanditiis beatae, amet, provident fuga hic! Nihil asperiores placeat modi sequi perferendis officia amet tenetur, error sit odio consequatur eligendi vero illum adipisci quia voluptatibus beatae vitae dolorem veniam ducimus!</p>
-                <p>Quibusdam, doloribus voluptate reiciendis provident assumenda unde maxime ullam a vitae nihil iusto ea amet ab tenetur sed dolor. Atque obcaecati perspiciatis magni. Magni ex dignissimos provident placeat cum tenetur modi distinctio beatae cupiditate, sit alias enim, cumque suscipit, minus obcaecati reiciendis rerum error? Magnam consectetur ut quisquam vero nemo!</p>
-            </div>
-        </article>
-    </section>
-</div>
+	<div class="example-wrapper">
+		<section class="figures">
+			<article>
+				<h2>Nom de la figure</h2>
+				<div class="metadata">Rédigé le 19/11/2021 à 15h00 dans la catégorie Figures</div>
+				<div class="content">
+					<img src="http://via.placeholder.com/350x150" alt="">
+					<p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reprehenderit facere illo nulla dolorem. Illo, ab.</p>
+					<p>Consequatur necessitatibus nostrum eaque. Omnis dolorum unde facilis eligendi odio quos quaerat debitis non quo!</p>
+					<hr>
+					<p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Itaque quo perspiciatis vero. Aliquam itaque aliquid, nostrum est dolor excepturi omnis blanditiis beatae, amet, provident fuga hic! Nihil asperiores placeat modi sequi perferendis officia amet tenetur, error sit odio consequatur eligendi vero illum adipisci quia voluptatibus beatae vitae dolorem veniam ducimus!</p>
+					<p>Quibusdam, doloribus voluptate reiciendis provident assumenda unde maxime ullam a vitae nihil iusto ea amet ab tenetur sed dolor. Atque obcaecati perspiciatis magni. Magni ex dignissimos provident placeat cum tenetur modi distinctio beatae cupiditate, sit alias enim, cumque suscipit, minus obcaecati reiciendis rerum error? Magnam consectetur ut quisquam vero nemo!</p>
+				</div>
+			</article>
+		</section>
+	</div>
 {% endblock %}
